docs(book): replace float with f32

+++ book/listings/ch03-structs/listing12.rs
@@ -10,8 +10,8 @@
 #             .spawn()
 #             .expect("Failed to spawn Runtime");
 #
-    let mut xy: StructRef = invoke_fn!(runtime, "vector2_new", -1.0f64, 1.0f64).unwrap();
-    let x: f64 = xy.get("x").unwrap();
+    let mut xy: StructRef = invoke_fn!(runtime, "vector2_new", -1.0f32, 1.0f32).unwrap();
+    let x: f32 = xy.get("x").unwrap();
     xy.set("x", x * x).unwrap();
-    let y = xy.replace("y", -1.0f64).unwrap();
+    let y = xy.replace("y", -1.0f32).unwrap();
 # }
